pyproject.toml
[project]
name = "book-sharing-app-friends"
version = "0.1.0"
description = ""
authors = [
    {name = "raistln",email = "samumarfon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13.2,<3.14.0"
dependencies = [
    "fastapi[all]>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.2",
    "pillow>=10.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "easyocr (>=1.7.2,<2.0.0)",
    "redis>=5.0.0",
    "pyzbar>=0.1.9",
    "opencv-python>=4.8.0",
    "slowapi>=0.1.9",
    "psutil>=5.9.0",
    "numpy>=1.26.0,<2.0.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
